refactor: add useControlPaneKeyboardShortcut
